Move function to create database items out of populatedb script and into its own module
